fonction zigzag inverse ok, fonction du print to ppm a l'air ok aussi, voir ce que demande la fonction

+++ src/color.c
@@ -1,5 +1,6 @@
 
 typedef struct RGB{
+    //Si l'image est en noir et blanc, on met le niveau de gris dans la composante R
     uint8_t R;
     uint8_t G;
     uint8_t B;
@@ -8,7 +9,7 @@ typedef struct RGB{
 
 public RGB ycbcr_to_rgb(float Y, float C_r, float C_b){
     RGB rgb;
-    int a = Y + 1.402*(C_r-128);
+    float a = Y + 1.402*(C_r-128);
     a = (r<0)?0:a;
     a = (r>255)?255:a;
     rgb.R = (uint_8t) a;
